fix: update indexGetter return type for Node 23 compatibility
Node 23 uses V8 version >12.4 which requires property interceptors to return
v8::Intercepted instead of void. Updated the indexGetter function to handle
both older and newer V8 versions by:
- Returning v8::Intercepted{} for V8 >12.4
- Maintaining void return for older versions
- Preserving the existing behavior of array access

The actual JavaScript return values remain unchanged, only the V8 interception
signaling mechanism is updated.

+++ src-cpp/utils.cpp
@@ -559,7 +559,7 @@ v8::Local<v8::Value> javaArrayToV8(Java *java, JNIEnv *env, jobjectArray objArra
     return v8::Int32Array::New(ab, 0, arraySize);
 #else
     for (jsize i = 0; i < arraySize; i++) {
-      result->Set(i, Nan::New<v8::Integer>(elems[i]));
+      result->Set(i, Nan::New<v8::Integer>(static_cast<int32_t>(elems[i])));
     }
     env->ReleaseIntArrayElements((jintArray)objArray, elems, 0);
 #endif
